Security fixes merge
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@4110 b456876b-0849-0410-b77d-98878d47e9d5

+++ tiki/tiki-backup.php
@@ -1,6 +1,6 @@
 <?php
 
-// $Header: /cvsroot/tikiwiki/tiki/tiki-backup.php,v 1.11 2004-03-28 07:32:23 mose Exp $
+// $Header: /cvsroot/tikiwiki/tiki/tiki-backup.php,v 1.12 2004-03-31 07:38:41 mose Exp $
 
 // Copyright (c) 2002-2004, Luis Argerich, Garland Foster, Eduardo Polidor, et. al.
 // All Rights Reserved. See copyright.txt for details and a complete list of authors.
@@ -36,15 +36,24 @@ if (isset($_REQUEST["restore"])) {
 }
 
 if (isset($_REQUEST["rrestore"])) {
-	check_ticket('backup');
-	$backuplib->restore_database("backups/$tikidomain" . basename($_REQUEST["rrestore"]));
+  $area = 'delbackup';
+  if (isset($_POST['daconfirm']) and isset($_SESSION["ticket_$area"])) {
+    key_check($area);
+		$backuplib->restore_database("backups/$tikidomain" . basename($_REQUEST["rrestore"]));
+	} else {
+		key_get($area);
+	}
 }
 
 if (isset($_REQUEST["remove"])) {
-	check_ticket('backup');
-	$filename = "backups/$tikidomain" . basename($_REQUEST["remove"]);
-
-	unlink ($filename);
+  $area = 'delbackup';
+  if (isset($_POST['daconfirm']) and isset($_SESSION["ticket_$area"])) {
+    key_check($area);
+		$filename = "backups/$tikidomain" . basename($_REQUEST["remove"]);
+		unlink ($filename);
+	} else {
+		key_get($area);
+	}
 }
 
 if (isset($_REQUEST["upload"])) {
